chore: move osinstall to cmd
