+++ cmd/osinstall/cmd/version.go
