+++ cmd/osinstall/internal/userdata/userdata.go
